pyproject.toml
[tool.poetry]
name = "pystarport"
version = "0.2.4"
description = "Spawn local devnets for cosmos-sdk chains"
authors = ["chain-dev-team <chain-dev@crypto.com>"]
license = "Apache"
keywords = ["CRO", "blockchain", "crypto.com", "cosmos"]
readme = "README.md"
homepage = "https://github.com/crypto-com/pystarport"
repository = "https://github.com/crypto-com/pystarport"

[tool.poetry.dependencies]
python = "^3.8"
fire = "^0.4.0"
tomlkit = "^0.7.0"
jsonmerge = "^1.7.0"
PyYAML = "^5.3.1"
python-dateutil = "^2.8.1"
durations = "^0.3.3"
supervisor = "^4.2.1"
docker = "^4.3.1"
bech32 = "^1.1.0"
multitail2 = "^1.5.2"
python-dotenv = "^0.19.2"
pyyaml-include = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
deepdiff = "^5.6.0"
flake8 = "^4.0.1"
black = "^21.12b0"
isort = "^5.10.1"

[tool.poetry.scripts]
pystarport = "pystarport.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
